[feat] enable speculative decoding for simple server (#113)
still have two issues: 
1> can't download both draft and target model in simple:
workaround would be one with name the other with path. for example:
```
/home/michael/code/ScaleLLM/build/src/server/simple --logtostderr --device=cuda:0 --model_name_or_path=baichuan-inc/Baichuan2-13B-Base --draft_model_name_or_path=/home/michael/.cache/huggingface/hub/models--baichuan-inc--Baichuan2-7B-Base/snapshots/f9d4d8dd2f7a3dbede3bda3b0cf0224e9272bbe5 --num_speculative_steps=3
```

2> rejection sampling result is not right

+++ src/request/sequence.h
@@ -113,7 +113,9 @@ class Sequence final {
   void append_new_token_id(int32_t next_token_id);
 
   // validate draft tokens with accepted tokens for speculative decoding
-  void validate_token_ids(const Slice<int32_t>& accpeted_token_ids);
+  // N.B. take int64_t as input to be compatible with torch::Tensor to avoid
+  // copy
+  void validate_token_ids(const Slice<int64_t>& accpeted_token_ids);
 
   // add new cache blocks
   void append_blocks(const std::vector<Block>& new_blocks);
